chatbot/requirements.txt
pymupdf
python-dotenv
sentence-transformers
faiss-cpu
tqdm
numpy
requests
fastapi
uvicorn
